moved tests for _resize_image to test_bbox
